pyproject.toml
[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[project]
name = "private-assistant-display-controller"
version = "0.0.1"
description = "Package is intended to be used in conjunction with the display skill to control a display device."
authors = [
    { name = "sktr22", email = "sktr22@github.com" }
]
readme = "README.md"
license = "GPL-3.0-only"
license-files = [
    "LICENSE",
]

keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
]

requires-python = ">=3.12,<3.14"
dependencies = [
    "aiomqtt>=2.0.0",
    "minio>=7.2.0",
    "Pillow>=10.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
rpi = ["inky>=2.0.0"]

[project.scripts]
inky-controller = "private_assistant_display_controller.main:app"

[project.urls]
Homepage = "https://github.com/stkr22/private-assistant-display-controller-py"
Repository = "https://github.com/stkr22/private-assistant-display-controller-py"
Documentation = "https://github.com/stkr22/private-assistant-display-controller-py/tree/main/docs"
Issues = "https://github.com/stkr22/private-assistant-display-controller-py/issues"
Changelog = "https://github.com/stkr22/private-assistant-display-controller-py/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "ruff~=0.14.1",
    "mypy~=1.17.0",
    "pytest~=8.4.1",
    "pytest-cov~=6.2.1",
    "pytest-xdist~=3.6.0",
    "coverage[toml]~=7.6.0",
    "pytest-asyncio~=0.25.2",
    "types-PyYAML>=6.0.0",
]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate
    "PL",   # pylint
    "RUF",  # ruff-specific rules
    "TCH",  # flake8-type-checking
]
ignore = ["E203"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]  # Magic values are acceptable in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "inky.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q -n auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["private_assistant_display_controller"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
]
